reorganize dir structure and make naming convention more logical
